decompositon/circlefit/convolution -> menpo.math
