--- menpo/decomposition/__init__.py
@@ -1,1 +1,0 @@
-from base import eigenvalue_decomposition, principal_component_decomposition
