+++ menpo/math/__init__.py
@@ -1,0 +1,4 @@
+from .circlefit import circle_fit
+from .convolution import log_gabor
+from .decomposition import (eigenvalue_decomposition,
+                            principal_component_decomposition)
