+++ menpo/math/circlefit.py
