+++ menpo/math/convolution.py
