--- menpo/misctools/__init__.py
@@ -1,1 +1,0 @@
-from menpo.misctools.circlefit import circle_fit
